+++ packages/string-find-heads-tails/changelog.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0] - 2017-12-28
+### Breaking API changes
+- ✨ Moved fourth argument, `fromIndex`, into `opts` because it stood in the way.
+
+### New
+- ✨ `opts.throwWhenSomethingWrongIsDetected` – When I tapped this API myself for the first time, I noticed errors with heads and tails should be `throw`n here, at this package, not at its consumers. For example, both heads and tails are found but they're in the opposite order. Now, the default settings will leads to `throw`n error. You can turn it off and make it behave like previous version by setting `opts.throwWhenSomethingWrongIsDetected` to `false`
+- ✨ `opts.allowWholeValueToBeOnlyHeadsOrTails` — When processing JSON data structures, it's possible that they will contain their own config. For example, JSON will use heads and tails, but also, there will be fields that DEFINE those heads and tails as well. This is a peculiar case - whole string will be equal to heads or tails. These cases will be recognised and errors won't be `throw`n. Unless you set `opts.allowWholeValueToBeOnlyHeadsOrTails` to `false`.
+- ✨ `optssource` - allows to `throw` errors in different name. Useful for parent libraries.
+
 ## [2.0.0] - 2017-12-27
 ### Changes
 Complete rewrite of the API.
@@ -42,4 +51,5 @@ now the same example in `v.2` API:
 ### New
 - ✨ First public release
 
+[3.0.0]: https://github.com/codsen/string-find-heads-tails/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/codsen/string-find-heads-tails/compare/v1.0.0...v2.0.0